+++ lib/Doctrine/ORM/Tools/Console/Command/ClearCache/QueryCommand.php
@@ -66,6 +66,10 @@ EOT
             throw new \InvalidArgumentException('No Query cache driver is configured on given EntityManager.');
         }
 
+        if ($cacheDriver instanceof \Doctrine\Common\Cache\ApcCache) {
+            throw new \LogicException("Cannot clear APC Cache from Console, its shared in the Webserver memory and not accessible from the CLI.");
+        }
+
         $output->write('Clearing ALL Query cache entries' . PHP_EOL);
 
         $cacheIds = $cacheDriver->deleteAll();
